# src/openapi.yaml
openapi: 3.1.0
info:
  title: Time-addressable Media Store
  description: 'The Time-addressable Media Store (TAMS) is used for storing segmented
    media Flows.

    The TAMS consists of one or more Storage Backends for the media Flow Segment Objects
    and a service providing a database index of the Flow Segments.

    This document is a specification of the service API.


    See the [bbc/tams](https://github.com/bbc/tams) repository for more background
    on TAMS.


    **Note**: the examples provided in this specification are for a system which uses
    a Storage Backend that provides HTTP URLs for uploading and downloading Media
    Objects in buckets.

    This could for example be implemented using an AWS S3 compatible store using presigned
    URLs or by a simple file system storage with an HTTP frontend.

    Clients should parse the [/service](#/operations/GET_service) endpoint and handle
    the Storage Backend type in use appropriately.

    '
  version: '8.0'
  contact:
    name: BBC R&D - Cloud-Fit Production Team
    email: cloudfit-opensource@rd.bbc.co.uk
    url: https://github.com/bbc/tams
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:4010
  description: Local mock of API
- url: https://example.com/tams/{version}
  description: Example TAMS service
  variables:
    version:
      description: API version
      default: v8.0
security:
- bearer_token_auth: []
- basic_auth: []
paths:
...
  /sources:
    head:
      summary: List Sources
      description: Return Sources path headers
      operationId: HEAD_sources
      tags:
      - Sources
      parameters:
      - name: label
        in: query
        description: Filter on Sources that have the given label.
        schema:
          type: string
      - name: tag.{name}
        in: query
        description: 'Filter on Sources that have a tag named {name} with a value
          in the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on Sources that have a tag named {name} regardless of
          value.

          {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - name: format
        in: query
        description: Filter on Source format.
        schema:
          $ref: vendor/bbc-tams/api/schemas/content-format.json
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/trait_resource_info_head_400'
    get:
      summary: List Sources
      description: List the Sources registered in the TAMS service instance and their
        details.
      operationId: GET_sources
      tags:
      - Sources
      parameters:
      - name: label
        in: query
        description: Filter on Sources that have the given label.
        schema:
          type: string
      - name: tag.{name}
        in: query
        description: 'Filter on Sources that have a tag named {name} with a value
          in the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on Sources that have a tag named {name} regardless of
          value.

          {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - name: format
        in: query
        description: Filter on Source format.
        schema:
          $ref: vendor/bbc-tams/api/schemas/content-format.json
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the service implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/sources-get-200.json
              schema:
                type: array
                items:
                  $ref: vendor/bbc-tams/api/schemas/source.json
        '400':
          description: Bad request. Invalid query options.
  /sources/{sourceId}:
    parameters:
    - name: sourceId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Source identifier.
    head:
      summary: Source Details
      description: Return Source headers
      operationId: HEAD_sources-sourceId
      tags:
      - Sources
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          $ref: '#/components/responses/trait_resource_info_head_404'
    get:
      summary: Source Details
      description: Returns Source metadata.
      operationId: GET_sources-sourceId
      tags:
      - Sources
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: vendor/bbc-tams/api/schemas/source.json
              examples:
                basic:
                  summary: Source containing elemental video
                  externalValue: vendor/bbc-tams/api/examples/source-get-200-basic.json
                multi:
                  summary: Multi-essence Source
                  description: Multi-essence Sources collect multiple Sources of different
                    formats under one Source ID.
                  externalValue: vendor/bbc-tams/api/examples/source-get-200-multi.json
        '404':
          description: The requested Source does not exist.
  /sources/{sourceId}/tags:
    parameters:
    - name: sourceId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Source identifier.
    head:
      summary: List Source Tags
      description: Return Source tags path headers
      operationId: HEAD_sources-sourceId-tags
      tags:
      - Sources
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_listing_head_200'
        '404':
          description: The requested Source does not exist.
    get:
      summary: List Source Tags
      description: Returns the Source tags.
      operationId: GET_sources-sourceId-tags
      tags:
      - Sources
      responses:
        '200':
          description: ''
          content:
            application/json:
              example:
                ingested_by: ingest_service_api
              schema:
                $ref: vendor/bbc-tams/api/schemas/tags.json
        '404':
          description: The requested Source does not exist.
  /sources/{sourceId}/tags/{name}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
      description: The tag name. {name} MUST be URL encoded where special characters
        are present.
    - name: sourceId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Source identifier.
    head:
      summary: Source Tag Value
      description: Return Source tag path headers
      operationId: HEAD_sources-sourceId-tags-name
      tags:
      - Sources
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          description: The requested Source or tag does not exist.
    get:
      summary: Source Tag Value
      description: Return the tag value associated with the tag name.
      operationId: GET_sources-sourceId-tags-name
      tags:
      - Sources
      responses:
        '200':
          description: ''
          content:
            application/json:
              example: '"ingest_service_api"

                '
              schema:
                type: string
        '404':
          description: The requested Source or tag does not exist.
    put:
      summary: Create or Update Source Tag
      description: Create or update the Source tag
      operationId: PUT_sources-sourceId-tags-name
      tags:
      - Sources
      requestBody:
        content:
          application/json:
            example: '"new_value"

              '
            schema:
              type: string
        required: true
      responses:
        '204':
          description: No content. The tag has been created or updated.
        '400':
          description: Bad request. Invalid Source tag value.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Source does not exist, or the tag name in the
            path is invalid.
    delete:
      summary: Delete Source Tag
      description: Delete a specific tag on a Source
      operationId: DELETE_sources-sourceId-tags-name
      tags:
      - Sources
      responses:
        '204':
          description: No content. The Source tag has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Source ID or tag in the path is invalid.
  /sources/{sourceId}/description:
    parameters:
    - name: sourceId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Source identifier.
    head:
      summary: Source Description
      description: Return Source description path headers
      operationId: HEAD_sources-sourceId-description
      tags:
      - Sources
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          $ref: '#/components/responses/trait_resource_info_head_404'
    get:
      summary: Source Description
      description: Returns the Source description property. This should be a human-readable
        description that may be showed in detailed views of Sources. The description
        should be longer and more detailed than `label`.
      operationId: GET_sources-sourceId-description
      tags:
      - Sources
      responses:
        '200':
          description: ''
          content:
            application/json:
              example: '"Big Buck Bunny"

                '
              schema:
                type: string
        '404':
          description: The requested Source does not exist.
    put:
      summary: Create or Update Source Description
      description: Create or update the description property. This should be a human-readable
        description that may be showed in detailed views of Sources. The description
        should be longer and more detailed than `label`.
      operationId: PUT_sources-sourceId-description
      tags:
      - Sources
      requestBody:
        content:
          application/json:
            example: '"Big Buck Bunny Movie"

              '
            schema:
              type: string
        required: true
      responses:
        '204':
          description: No content. The description has been created or updated.
        '400':
          description: Bad request. Invalid Source description.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Source does not exist.
    delete:
      summary: Delete Source Description
      description: Delete the description property.
      operationId: DELETE_sources-sourceId-description
      tags:
      - Sources
      responses:
        '204':
          description: No content. The Source description property has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The Source ID in the path is invalid.
  /sources/{sourceId}/label:
    parameters:
    - name: sourceId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Source identifier.
    head:
      summary: Source Label
      description: Return Source label path headers
      operationId: HEAD_sources-sourceId-label
      tags:
      - Sources
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_info_head_200'
        '404':
          description: The requested Source does not exist, or does not have a label
            set.
    get:
      summary: Source Label
      description: Returns the Source label property. This should be a very short,
        human-readable label that may be displayed in listings of Sources.
      operationId: GET_sources-sourceId-label
      tags:
      - Sources
      responses:
        '200':
          description: ''
          content:
            application/json:
              example: '"Big Buck Bunny"

                '
              schema:
                type: string
        '404':
          description: The requested Source does not exist, or does not have a label
            set.
    put:
      summary: Create or Update Source Label
      description: Create or update the label property. This should be a very short,
        human-readable label that may be displayed in listings of Sources.
      operationId: PUT_sources-sourceId-label
      tags:
      - Sources
      requestBody:
        content:
          application/json:
            example: '"Big Buck Bunny Movie"

              '
            schema:
              type: string
        required: true
      responses:
        '204':
          description: No content. The label has been created or updated.
        '400':
          description: Bad request. Invalid Source label.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Source does not exist.
    delete:
      summary: Delete Source Label
      description: Delete the label property.
      operationId: DELETE_sources-sourceId-label
      tags:
      - Sources
      responses:
        '204':
          description: No content. The Source label property has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Source ID in the path is invalid.
...
components:
  responses:
    trait_resource_listing_head_200:
      description: ''
      content:
        application/json:
          schema:
            type: string
    trait_resource_info_head_200:
      description: ''
      content:
        application/json:
          schema:
            type: string
    trait_resource_info_head_400:
      description: Bad request. Query parameters are invalid.
    trait_resource_info_head_404:
      description: Resource was not found.
  parameters:
    trait_paged_limit:
      name: limit
      in: query
      description: Restrict the response to the specified number of results. Service
        implementations may specify their own default and maximum for the limit
      schema:
        type: integer
    trait_resource_paged_key:
      name: page
      in: query
      description: Opaque string used by backend to access a specific page of results.
        Clients should read the next URL from the `Link` header returned with responses,
        or use value of the returned X-Paging-NextKey header. If not supplied, the
        first page is accessed. Service implementations should ensure a consistent
        sort order is applied to pages of results.
      schema:
        type: string
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
      description: 'HTTP basic authentication using a username and password.


        This option should only be used in constrained cases where Bearer tokens are
        impractical, since the password is intrinsically long-lived, and managing
        user credentials across systems is difficult compared to using centralised
        auth (e.g. via OAuth2) with Bearer tokens

        '
    bearer_token_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HTTP bearer token authentication.


        The method through which a client should acquire a bearer token is assumed
        to be an OAuth2 grant (either the Authorization Code or Client Credentials
        grant), however the OpenAPI `OAuth Flow` object cannot be used because it
        requires that the relevant URLs are known in advance.


        Clients should provide options to allow one (or both) grant types to be configured,
        as appropriate.


        Where tokens are not JWTs, or are acquired through some other mechanism, the
        API key method should be used instead.

        '
...